Read ability damage as double
- Read EarthLine, MetalFragments, and IceWall damage as double.
- Change default config values to double version to indicate that they
  don't need to be integers.
- Increase version to 2.4.3.

+++ src/com/jedk1/jedcore/ability/earthbending/EarthLine.java
@@ -61,7 +61,7 @@ public class EarthLine extends EarthAbility implements AddonAbility {
 		range = JedCore.plugin.getConfig().getInt("Abilities.Earth.EarthLine.Range");
 		preparerange = JedCore.plugin.getConfig().getInt("Abilities.Earth.EarthLine.PrepareRange");
 		affectingradius = JedCore.plugin.getConfig().getInt("Abilities.Earth.EarthLine.AffectingRadius");
-		damage = JedCore.plugin.getConfig().getInt("Abilities.Earth.EarthLine.Damage");
+		damage = JedCore.plugin.getConfig().getDouble("Abilities.Earth.EarthLine.Damage");
 	}
 
 	public boolean prepare() {
